@@ -9,7 +9,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@chakra-ui/react": "^2.1.2",
+    "@chakra-ui/theme-tools": "^2.0.1",
     "@coinbase/wallet-sdk": "^3.0.5",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
     "@lottiefiles/react-lottie-player": "^3.4.2",
     "@near-wallet-selector/core": "^3.1.0",
     "@near-wallet-selector/near-wallet": "^3.1.0",
@@ -35,7 +39,7 @@
     "crypto-js": "^4.1.1",
     "ethers": "^5.6.1",
     "final-form": "^4.20.6",
-    "framer-motion": "^6.3.5",
+    "framer-motion": "^6",
     "lodash": "^4.17.21",
     "mobile-detect": "^1.4.5",
     "moment-timezone": "^0.5.34",
@@ -43,14 +47,15 @@
     "near-api-js": "^0.44.2",
     "query-string": "^7.1.1",
     "rc-switch": "^3.2.2",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-bootstrap": "^2.1.2",
     "react-content-loader": "^6.1.0",
     "react-copy-to-clipboard": "^5.0.4",
     "react-custom-scrollbars-2": "^4.4.0",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-final-form": "^6.5.8",
+    "react-icons": "^4.4.0",
     "react-loading-overlay": "^1.0.1",
     "react-redux": "^7.2.6",
     "react-reveal": "^1.2.2",
